change compilor to c++

+++ ex00/megaphone.cpp
@@ -11,7 +11,7 @@ int main(int argc, char **argv) {
 	}
 	for (int i = 1; i < argc; i++) {
 		std::string str(argv[i]);
-		for (int j = 0; j < str.size(); j++) {
+		for (size_t j = 0; j < str.size(); j++) {
 			str[j] = toupper(str[j]);
 		}
 		result += str;
